H2 and Spring Tests
- change db to h2
- add spring tests

+++ src/main/java/br/com/zupacademy/alonso/casadocodigo/repository/AuthorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import br.com.zupacademy.alonso.casadocodigo.model.Author;
 
 @Repository
-public interface AuthorRepository extends CrudRepository<Author,Integer>{
+public interface AuthorRepository extends CrudRepository<Author,Long>{
 	//Optional<Author> findByEmail(String email);
 	boolean existsByEmail(String email);
 }
